Rendered thumbnails based off of a collections results from the REST api
There was a problem that took a long time to debug where the
Backbone.History.start() wasn't the first thing being called. This
caused some very cryptic error messages. To fix it, I had to call the
start() method from within the base router and then require the base
view. I think this happened based on how the generator structed things
to work in both development, test and produciton.